refactor: extract pixelformat conversions to different file
